services/backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
aerich==0.7.1
asyncpg==0.27.0
fastapi==0.88.0
tortoise-orm==0.19.2
uvicorn==0.20.0
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
